+++ src/room/RTCEngine.ts
@@ -17,6 +17,7 @@ import {
   AddTrackRequest,
   JoinResponse,
   LeaveRequest,
+  ReconnectResponse,
   SignalTarget,
   TrackPublishedResponse,
 } from '../proto/livekit_rtc';
@@ -29,7 +30,7 @@ import {
   UnexpectedConnectionState,
 } from './errors';
 import { EngineEvent } from './events';
-import PCTransport from './PCTransport';
+import PCTransport, { PCEvents } from './PCTransport';
 import type { ReconnectContext, ReconnectPolicy } from './ReconnectPolicy';
 import type LocalTrack from './track/LocalTrack';
 import type LocalVideoTrack from './track/LocalVideoTrack';
@@ -185,8 +186,13 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
 
   async close() {
     const unlock = await this.closingLock.lock();
+    if (this.isClosed) {
+      unlock();
+      return;
+    }
     try {
       this._isClosed = true;
+      this.emit(EngineEvent.Closing);
       this.removeAllListeners();
       this.deregisterOnLineListener();
       this.clearPendingReconnect();
@@ -274,35 +280,7 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
 
     this.participantSid = joinResponse.participant?.sid;
 
-    const rtcConfig = { ...this.rtcConfig };
-
-    // update ICE servers before creating PeerConnection
-    if (joinResponse.iceServers && !rtcConfig.iceServers) {
-      const rtcIceServers: RTCIceServer[] = [];
-      joinResponse.iceServers.forEach((iceServer) => {
-        const rtcIceServer: RTCIceServer = {
-          urls: iceServer.urls,
-        };
-        if (iceServer.username) rtcIceServer.username = iceServer.username;
-        if (iceServer.credential) {
-          rtcIceServer.credential = iceServer.credential;
-        }
-        rtcIceServers.push(rtcIceServer);
-      });
-      rtcConfig.iceServers = rtcIceServers;
-    }
-
-    if (
-      joinResponse.clientConfiguration &&
-      joinResponse.clientConfiguration.forceRelay === ClientConfigSetting.ENABLED
-    ) {
-      rtcConfig.iceTransportPolicy = 'relay';
-    }
-
-    // @ts-ignore
-    rtcConfig.sdpSemantics = 'unified-plan';
-    // @ts-ignore
-    rtcConfig.continualGatheringPolicy = 'gather_continually';
+    const rtcConfig = this.makeRTCConfiguration(joinResponse);
 
     if (this.signalOpts?.e2eeEnabled) {
       log.debug('E2EE - setting up transports with insertable streams');
@@ -448,6 +426,40 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
       }
       log.trace('leave request', { leave });
     };
+  }
+
+  private makeRTCConfiguration(serverResponse: JoinResponse | ReconnectResponse): RTCConfiguration {
+    const rtcConfig = { ...this.rtcConfig };
+
+    // update ICE servers before creating PeerConnection
+    if (serverResponse.iceServers && !rtcConfig.iceServers) {
+      const rtcIceServers: RTCIceServer[] = [];
+      serverResponse.iceServers.forEach((iceServer) => {
+        const rtcIceServer: RTCIceServer = {
+          urls: iceServer.urls,
+        };
+        if (iceServer.username) rtcIceServer.username = iceServer.username;
+        if (iceServer.credential) {
+          rtcIceServer.credential = iceServer.credential;
+        }
+        rtcIceServers.push(rtcIceServer);
+      });
+      rtcConfig.iceServers = rtcIceServers;
+    }
+
+    if (
+      serverResponse.clientConfiguration &&
+      serverResponse.clientConfiguration.forceRelay === ClientConfigSetting.ENABLED
+    ) {
+      rtcConfig.iceTransportPolicy = 'relay';
+    }
+
+    // @ts-ignore
+    rtcConfig.sdpSemantics = 'unified-plan';
+    // @ts-ignore
+    rtcConfig.continualGatheringPolicy = 'gather_continually';
+
+    return rtcConfig;
   }
 
   private createDataChannels() {
@@ -851,7 +863,12 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     }
 
     try {
-      await this.client.reconnect(this.url, this.token, this.participantSid);
+      const res = await this.client.reconnect(this.url, this.token, this.participantSid);
+      if (res) {
+        const rtcConfig = this.makeRTCConfiguration(res);
+        this.publisher.pc.setConfiguration(rtcConfig);
+        this.subscriber.pc.setConfiguration(rtcConfig);
+      }
     } catch (e) {
       let message = '';
       if (e instanceof Error) {
@@ -930,52 +947,99 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     }
   }
 
-  private async ensurePublisherConnected(kind: DataPacket_Kind) {
-    if (!this.subscriberPrimary) {
-      return;
+  /**
+   * @internal
+   */
+  async ensureDataTransportConnected(
+    kind: DataPacket_Kind,
+    subscriber: boolean = this.subscriberPrimary,
+  ) {
+    const transport = subscriber ? this.subscriber : this.publisher;
+    const transportName = subscriber ? 'Subscriber' : 'Publisher';
+    if (!transport) {
+      throw new ConnectionError(`${transportName} connection not set`);
     }
 
-    if (!this.publisher) {
-      throw new ConnectionError('publisher connection not set');
-    }
-
-    if (!this.publisher.isICEConnected && this.publisher.pc.iceConnectionState !== 'checking') {
+    if (
+      !subscriber &&
+      !this.publisher?.isICEConnected &&
+      this.publisher?.pc.iceConnectionState !== 'checking'
+    ) {
       // start negotiation
       this.negotiate();
     }
 
-    const targetChannel = this.dataChannelForKind(kind);
+    const targetChannel = this.dataChannelForKind(kind, subscriber);
     if (targetChannel?.readyState === 'open') {
       return;
     }
 
-    // wait until publisher ICE connected
+    // wait until ICE connected
     const endTime = new Date().getTime() + this.peerConnectionTimeout;
     while (new Date().getTime() < endTime) {
-      if (this.publisher.isICEConnected && this.dataChannelForKind(kind)?.readyState === 'open') {
+      if (
+        transport.isICEConnected &&
+        this.dataChannelForKind(kind, subscriber)?.readyState === 'open'
+      ) {
         return;
       }
       await sleep(50);
     }
 
     throw new ConnectionError(
-      `could not establish publisher connection, state ${this.publisher?.pc.iceConnectionState}`,
+      `could not establish ${transportName} connection, state: ${transport.pc.iceConnectionState}`,
     );
   }
 
+  private async ensurePublisherConnected(kind: DataPacket_Kind) {
+    await this.ensureDataTransportConnected(kind, false);
+  }
+
   /** @internal */
-  negotiate() {
-    if (!this.publisher) {
-      return;
-    }
-
-    this.hasPublished = true;
-
-    this.publisher.negotiate((e) => {
-      if (e instanceof NegotiationError) {
-        this.fullReconnectOnNext = true;
+  negotiate(): Promise<void> {
+    // observe signal state
+    return new Promise<void>((resolve, reject) => {
+      if (!this.publisher) {
+        reject(new NegotiationError('publisher is not defined'));
+        return;
       }
-      this.handleDisconnect('negotiation');
+
+      this.hasPublished = true;
+
+      const handleClosed = () => {
+        log.debug('engine disconnected while negotiation was ongoing');
+        cleanup();
+        resolve();
+        return;
+      };
+
+      this.on(EngineEvent.Closing, handleClosed);
+
+      const negotiationTimeout = setTimeout(() => {
+        reject('negotiation timed out');
+        this.handleDisconnect('negotiation');
+      }, this.peerConnectionTimeout);
+
+      const cleanup = () => {
+        clearTimeout(negotiationTimeout);
+        this.off(EngineEvent.Closing, handleClosed);
+      };
+
+      this.publisher.once(PCEvents.NegotiationStarted, () => {
+        this.publisher?.once(PCEvents.NegotiationComplete, () => {
+          cleanup();
+          resolve();
+        });
+      });
+
+      this.publisher.negotiate((e) => {
+        cleanup();
+        reject(e);
+        if (e instanceof NegotiationError) {
+          this.fullReconnectOnNext = true;
+        }
+        this.handleDisconnect('negotiation');
+      });
     });
   }
 
@@ -1071,6 +1135,7 @@ export type EngineEventCallbacks = {
   restarting: () => void;
   restarted: (joinResp: JoinResponse) => void;
   signalResumed: () => void;
+  closing: () => void;
   mediaTrackAdded: (
     track: MediaStreamTrack,
     streams: MediaStream,